requirements.txt
langchain>=0.0.200
openai>=0.27.0
python-dotenv>=0.19.0
dnspython>=2.3.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-nmap>=0.7.1
aiohttp>=3.8.0
pydantic>=2.0.0
rich>=13.0.0
wfuzz>=3.1.0
wafw00f>=2.2.0
reportlab>=4.0.0